Delete ebpf object files
We no longer need to store files in git history.
We used to need it to simplify development. Most developers don't edit
the ebpf programs and we wanted to avoid having everyone having to set
up a build environment.
Now we have the mizu-ebpf-builder docker which is an easy way to compile
ebpf.
When building the agent image, we used to take the local object files
from the working dir. Now the objects are compiled when building the
docker image.